+++ src/Schema/Response/Picture.php
@@ -4,6 +4,7 @@ namespace MssPhp\Schema\Response;
 
 use JMS\Serializer\Annotation\AccessType;
 use JMS\Serializer\Annotation\Type;
+use JMS\Serializer\Annotation\PostDeserialize;
 
 class Picture
 {
@@ -36,4 +37,16 @@ class Picture
      * @Type("integer")
      */
     public $height;
+
+    /**
+     * @PostDeserialize
+     */
+    public function postDeserialize()
+    {
+        $this->url = str_replace(
+            "https://easychannel.it/",
+            "https://cdn.easychannel.it/",
+            $this->url
+        );
+    }
 }