+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Update dependencies
+- Rewrite picture URLs to https://cdn.easychannel.it/…
 
 ## [6.1.0] - 2024-10-16